add documentation for identity_derive.
